requirements.txt
streamlit==1.40.2
snowflake-connector-python==3.12.3
pandas==2.2.3
